Refactor ghost replay logic and implement Fast Reset
- Combined redundant `process_course_ghost_replay` and `process_player_ghost_replay` in `src/replays.c` into a single `process_ghost_replay_generic` function.
- Upgraded the Level Reset (L+R+Start) feature in `src/main.c` to perform a "Fast Reset". It now instantly resets state variables and warps the player to the start line without forcing a full course reload.
- Fixed UI text rendering and enum missing for the "Save Ghost" Practice option in `src/menu_items.c`.
- Added missing PC Port abstraction header `include/pc_port/platform_save.h`.
- Addressed code review feedback: pointer type in replays.c and missing return in menus.c.
- Updated `MANUAL.md` and `CHANGELOG.md` to reflect v1.6.1 changes.

+++ src/replays.c
@@ -324,7 +324,7 @@ void process_post_time_trial_replay(void) {
     }
 }
 
-void process_ghost_replay_generic(u32* replayData, s32* replayIdx, s32* framesRemaining, u16* buttonsPrev, struct Controller* controller, Player* playerToStop) {
+void process_ghost_replay_generic(u32* replayData, s16* replayIdx, u32* framesRemaining, u16* buttonsPrev, struct Controller* controller, Player* playerToStop) {
     u32 inputs;
     u32 stickBytes;
     u16 buttonsTemp;
